vt - finished first couple of unit tests
