+++ jest.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  transform: {
+    "^.+\\.jsx?$": "babel-jest",
+  },
+  testPathIgnorePatterns: ["/node_modules/", "/cypress/"],
+};
