+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
+    "test:jest": "jest",
     "dev": "next dev",
     "build": "next build",
     "start": "next start -p $PORT",
@@ -37,6 +38,7 @@
     "terser-webpack-plugin": "^2.3.5"
   },
   "devDependencies": {
+    "jest": "^25.4.0",
     "@rckeller/cypress-unfetch": "^1.0.1",
     "cypress": "^4.2.0",
     "danger": "^10.0.0",
